DEVPROD-18765: Remove legacy Task page
