+++ config_serviceflags.go
@@ -39,7 +39,6 @@ type ServiceFlags struct {
 	ElasticIPsDisabled              bool `bson:"elastic_ips_disabled" json:"elastic_ips_disabled"`
 	ReleaseModeDisabled             bool `bson:"release_mode_disabled" json:"release_mode_disabled"`
 	AdminParameterStoreDisabled     bool `bson:"admin_parameter_store_disabled" json:"admin_parameter_store_disabled"`
-	LegacyUITaskPageDisabled        bool `bson:"legacy_ui_task_page_disabled" json:"legacy_ui_task_page_disabled"`
 	LegacyUITaskHistoryPageDisabled bool `bson:"legacy_ui_task_history_page_disabled" json:"legacy_ui_task_history_page_disabled"`
 
 	// Notification Flags
@@ -96,7 +95,6 @@ func (c *ServiceFlags) Set(ctx context.Context) error {
 			cpuDegradedModeDisabledKey:         c.CPUDegradedModeDisabled,
 			releaseModeDisabledKey:             c.ReleaseModeDisabled,
 			adminParameterStoreDisabledKey:     c.AdminParameterStoreDisabled,
-			legacyUITaskPageDisabledKey:        c.LegacyUITaskPageDisabled,
 			legacyUITaskHistoryPageDisabledKey: c.LegacyUITaskHistoryPageDisabled,
 		}}), "updating config section '%s'", c.SectionId(),
 	)
