+++ public/static/js/directives/directives.eventlogs.js
@@ -2,18 +2,6 @@ var directives = directives || {};
 
 directives.eventlogs = angular.module('directives.eventLogs', ['filters.common']);
 
-directives.eventlogs.directive('taskevent', function() {
-  return {
-    scope:{
-      userTz:"=tz",
-      eventLogObj:"=event",
-      jira:"=jira"
-    },
-    restrict : 'E',
-    templateUrl : '/static/partials/taskevent.html',
-  };
-});
-
 directives.eventlogs.directive('hostevent', function() {
   return {
     scope:{
