--- public/static/partials/taskevent.html
@@ -1,19 +1,0 @@
-<div class="eventlog">
-  <div class="timestamp">[[eventLogObj.timestamp | convertDateToUserTimezone:userTz:'MMM D, YYYY h:mm:ss a']]</div>
-  <div class="event_details" ng-switch="eventLogObj.event_type">
-    <span ng-switch-when="TASK_FINISHED">Completed with status: <b>[[eventLogObj.data.status]]</b></span>
-    <span ng-switch-when="TASK_STARTED">Marked as "started"</span>
-    <span ng-switch-when="TASK_DISPATCHED">Dispatched to host <a href="/host/[[eventLogObj.data.host_id]]">[[eventLogObj.data.host_id]]</a></span>
-    <span ng-switch-when="TASK_UNDISPATCHED">Undispatched from host <a href="/host/[[eventLogObj.data.host_id]]">[[eventLogObj.data.host_id]]</a></span>
-    <span ng-switch-when="TASK_CREATED">Task created</span>
-    <span ng-switch-when="TASK_RESTARTED">Restarted by [[eventLogObj.data.user_id]].</span>
-    <span ng-switch-when="TASK_ACTIVATED">Activated by [[eventLogObj.data.user_id]].</span>
-    <span ng-switch-when="TASK_JIRA_ALERT_CREATED">Created Jira Alert <strong ng-bind-html="eventLogObj.data.jira | jiraLinkify: jira | ansi"></strong>.</span>
-    <span ng-switch-when="TASK_DEACTIVATED">Deactivated by user [[eventLogObj.data.user_id]].</span>
-    <span ng-switch-when="TASK_ABORT_REQUEST">Marked to abort by user [[eventLogObj.data.user_id]].</span>
-    <span ng-switch-when="TASK_PRIORITY_CHANGED">Priority Changed to [[eventLogObj.data.priority]] by [[eventLogObj.data.user_id]]</span>
-    <span ng-switch-when="TASK_DEPENDENCIES_OVERRIDDEN">Dependencies overridden by user [[eventLogObj.data.user_id]].</span>
-    <span ng-switch-when="MERGE_TASK_UNSCHEDULED">Merge Task Unscheduled by user [[eventLogObj.data.user_id]].</span>
-  </div>
-  <div class="clearfix"></div>
-</div>
