+++ rest/model/admin.go
@@ -2186,7 +2186,6 @@ type APIServiceFlags struct {
 	ElasticIPsDisabled              bool `json:"elastic_ips_disabled"`
 	ReleaseModeDisabled             bool `json:"release_mode_disabled"`
 	AdminParameterStoreDisabled     bool `json:"admin_parameter_store_disabled"`
-	LegacyUITaskPageDisabled        bool `json:"legacy_ui_task_page_disabled"`
 	LegacyUITaskHistoryPageDisabled bool `json:"legacy_ui_task_history_page_disabled"`
 
 	// Notifications Flags
@@ -2619,7 +2618,6 @@ func (as *APIServiceFlags) BuildFromService(h any) error {
 		as.ElasticIPsDisabled = v.ElasticIPsDisabled
 		as.ReleaseModeDisabled = v.ReleaseModeDisabled
 		as.AdminParameterStoreDisabled = v.AdminParameterStoreDisabled
-		as.LegacyUITaskPageDisabled = v.LegacyUITaskPageDisabled
 		as.LegacyUITaskHistoryPageDisabled = v.LegacyUITaskHistoryPageDisabled
 	default:
 		return errors.Errorf("programmatic error: expected service flags config but got type %T", h)
@@ -2666,7 +2664,6 @@ func (as *APIServiceFlags) ToService() (any, error) {
 		ElasticIPsDisabled:              as.ElasticIPsDisabled,
 		ReleaseModeDisabled:             as.ReleaseModeDisabled,
 		AdminParameterStoreDisabled:     as.AdminParameterStoreDisabled,
-		LegacyUITaskPageDisabled:        as.LegacyUITaskPageDisabled,
 		LegacyUITaskHistoryPageDisabled: as.LegacyUITaskHistoryPageDisabled,
 	}, nil
 }
