+++ service/task.go
@@ -1,7 +1,6 @@
 package service
 
 import (
-	"context"
 	"encoding/json"
 	"fmt"
 	"io"
@@ -14,146 +13,25 @@ import (
 	"github.com/evergreen-ci/evergreen/apimodels"
 	"github.com/evergreen-ci/evergreen/model"
 	"github.com/evergreen-ci/evergreen/model/artifact"
-	"github.com/evergreen-ci/evergreen/model/build"
-	"github.com/evergreen-ci/evergreen/model/distro"
 	"github.com/evergreen-ci/evergreen/model/event"
-	"github.com/evergreen-ci/evergreen/model/host"
 	"github.com/evergreen-ci/evergreen/model/log"
 	"github.com/evergreen-ci/evergreen/model/task"
-	"github.com/evergreen-ci/evergreen/model/testresult"
 	"github.com/evergreen-ci/evergreen/model/user"
-	"github.com/evergreen-ci/evergreen/plugin"
 	"github.com/evergreen-ci/gimlet"
-	"github.com/evergreen-ci/gimlet/rolemanager"
 	"github.com/evergreen-ci/utility"
-	"github.com/mongodb/grip"
-	"github.com/mongodb/grip/message"
 	"github.com/pkg/errors"
-	"go.mongodb.org/mongo-driver/bson"
 )
-
-type uiTaskData struct {
-	Id                   string                  `json:"id"`
-	DisplayName          string                  `json:"display_name"`
-	Revision             string                  `json:"gitspec"`
-	BuildVariant         string                  `json:"build_variant"`
-	Distro               string                  `json:"distro"`
-	Container            string                  `json:"container,omitempty"`
-	BuildId              string                  `json:"build_id"`
-	Status               string                  `json:"status"`
-	TaskWaiting          string                  `json:"task_waiting"`
-	Activated            bool                    `json:"activated"`
-	Execution            int                     `json:"execution"`
-	TotalExecutions      int                     `json:"total_executions"`
-	StartTime            int64                   `json:"start_time"`
-	DispatchTime         int64                   `json:"dispatch_time"`
-	FinishTime           int64                   `json:"finish_time"`
-	Requester            string                  `json:"r"`
-	ExpectedDuration     time.Duration           `json:"expected_duration"`
-	Priority             int64                   `json:"priority"`
-	TimeTaken            time.Duration           `json:"time_taken"`
-	TaskEndDetails       apimodels.TaskEndDetail `json:"task_end_details"`
-	TestResults          []uiTestResult          `json:"test_results"`
-	Aborted              bool                    `json:"abort"`
-	AbortInfo            task.AbortInfo          `json:"abort_info"`
-	MinQueuePos          int                     `json:"min_queue_pos"`
-	DependsOn            []uiDep                 `json:"depends_on"`
-	AbortedByDisplay     *abortedByDisplay       `json:"aborted_by_display,omitempty"`
-	OverrideDependencies bool                    `json:"override_dependencies"`
-	IngestTime           time.Time               `json:"ingest_time"`
-	EstWaitTime          time.Duration           `json:"wait_time"`
-	UpstreamData         *uiUpstreamData         `json:"upstream_data,omitempty"`
-
-	// from the host doc (the dns name)
-	HostDNS string `json:"host_dns,omitempty"`
-	// from the host doc (the host id)
-	HostId string `json:"host_id,omitempty"`
-
-	// for breadcrumb
-	BuildVariantDisplay string `json:"build_variant_display"`
-
-	// from version
-	VersionId   string    `json:"version_id"`
-	Message     string    `json:"message"`
-	Project     string    `json:"branch"`
-	Author      string    `json:"author"`
-	AuthorEmail string    `json:"author_email"`
-	CreateTime  time.Time `json:"create_time"`
-
-	// from project
-	RepoOwner string `json:"repo_owner"`
-	Repo      string `json:"repo_name"`
-
-	// to avoid time skew b/t browser and API server
-	CurrentTime int64 `json:"current_time"`
-
-	// flag to indicate whether this is the current execution of this task, or
-	// a previous execution
-	Archived bool `json:"archived"`
-
-	PatchInfo *uiPatch `json:"patch_info"`
-
-	// display task info
-	DisplayOnly    bool         `json:"display_only"`
-	ExecutionTasks []uiExecTask `json:"execution_tasks"`
-	PartOfDisplay  bool         `json:"in_display"`
-	DisplayTaskID  string       `json:"display_task,omitempty"`
-
-	// generated task info
-	GeneratedById   string `json:"generated_by_id"`
-	GeneratedByName string `json:"generated_by_name"`
-}
-
-type uiDep struct {
-	Id             string                  `json:"id"`
-	Name           string                  `json:"display_name"`
-	Status         string                  `json:"status"`
-	RequiredStatus string                  `json:"required"`
-	Activated      bool                    `json:"activated"`
-	BuildVariant   string                  `json:"build_variant"`
-	Details        apimodels.TaskEndDetail `json:"task_end_details"`
-	Recursive      bool                    `json:"recursive"`
-	TaskWaiting    string                  `json:"task_waiting"`
-}
-
-type uiExecTask struct {
-	Id        string        `json:"id"`
-	Name      string        `json:"display_name"`
-	Status    string        `json:"status"`
-	TimeTaken time.Duration `json:"time_taken"`
-}
-
-type uiTestResult struct {
-	TestResult testresult.TestResult `json:"test_result"`
-	TaskId     string                `json:"task_id"`
-	TaskName   string                `json:"task_name"`
-	URL        string                `json:"url"`
-	URLRaw     string                `json:"url_raw"`
-	URLParsley string                `json:"url_parsley"`
-}
 
 type logData struct {
 	Data chan apimodels.LogMessage
 	User gimlet.User
 }
 
-type abortedByDisplay struct {
-	TaskDisplayName     string `json:"task_display_name"`
-	BuildVariantDisplay string `json:"build_variant_display"`
-}
-
 func (uis *UIServer) taskPage(w http.ResponseWriter, r *http.Request) {
-	ctx := r.Context()
-
-	flags, err := evergreen.GetServiceFlags(r.Context())
-	if err != nil {
-		gimlet.WriteResponse(w, gimlet.MakeJSONInternalErrorResponder(errors.Wrap(err, "retrieving admin settings")))
-		return
-	}
-
 	projCtx := MustHaveProjectContext(r)
 	executionStr := gimlet.GetVars(r)["execution"]
 	var execution int
+	var err error
 
 	if executionStr != "" {
 		execution, err = strconv.Atoi(executionStr)
@@ -162,345 +40,16 @@ func (uis *UIServer) taskPage(w http.ResponseWriter, r *http.Request) {
 			return
 		}
 	}
-
 	if projCtx.Task == nil {
 		http.Error(w, "Not found", http.StatusNotFound)
 		return
 	}
-
 	spruceLink := fmt.Sprintf("%s/task/%s?execution=%d", uis.Settings.Ui.UIv2Url, projCtx.Task.Id, execution)
-
-	// TODO: Delete all content in this file with the exception of this redirect.
-	if flags.LegacyUITaskPageDisabled {
-		http.Redirect(w, r, spruceLink, http.StatusPermanentRedirect)
-		return
-	}
-
-	if RedirectSpruceUsers(w, r, spruceLink) {
-		return
-	}
-
-	if projCtx.Build == nil {
-		uis.LoggedError(w, r, http.StatusInternalServerError, errors.New("build not found"))
-		return
-	}
-
-	if projCtx.Version == nil {
-		uis.LoggedError(w, r, http.StatusInternalServerError, errors.New("version not found"))
-		return
-	}
-
-	if projCtx.ProjectRef == nil {
-		grip.Error("Project ref is nil")
-		uis.LoggedError(w, r, http.StatusInternalServerError, errors.New("version not found"))
-		return
-	}
-
-	archived := false
-
-	// if there is an execution number, the task might be in the old_tasks collection, so we
-	// query that collection and set projCtx.Task to the old task if it exists.
-	if executionStr != "" {
-		// Construct the old task id.
-		oldTaskId := task.MakeOldID(projCtx.Task.Id, execution)
-
-		// Try to find the task in the old_tasks collection.
-		var taskFromDb *task.Task
-		taskFromDb, err = task.FindOneOld(r.Context(), task.ById(oldTaskId))
-		if err != nil {
-			uis.LoggedError(w, r, http.StatusInternalServerError, err)
-			return
-		}
-
-		// If we found a task, set the task context. Otherwise, if taskFromDb is nil, check
-		// that the execution matches the context's execution. If it does not, return an
-		// error, since that means we are searching for a task that does not exist.
-		if taskFromDb != nil {
-			projCtx.Task = taskFromDb
-			archived = true
-		} else if execution != projCtx.Task.Execution {
-			uis.LoggedError(w, r, http.StatusNotFound, errors.New("Error finding task or execution"))
-			return
-		}
-	}
-
-	// Build a struct containing the subset of task data needed for display in the UI
-	tId := projCtx.Task.Id
-	totalExecutions := projCtx.Task.Execution
-	taskExecution := projCtx.Task.Execution
-	if archived {
-		tId = projCtx.Task.OldTaskId
-
-		// Get total number of executions for executions drop down
-		mostRecentExecution, err := task.FindOneId(r.Context(), tId)
-		if err != nil {
-			uis.LoggedError(w, r, http.StatusInternalServerError,
-				errors.Wrapf(err, "Error finding most recent execution by id %s", tId))
-			return
-		}
-		totalExecutions = mostRecentExecution.Execution
-	}
-	if totalExecutions < 1 {
-		totalExecutions = 1
-	}
-
-	uiTask := uiTaskData{
-		Id:                   tId,
-		DisplayName:          projCtx.Task.DisplayName,
-		Revision:             projCtx.Task.Revision,
-		Status:               projCtx.Task.Status,
-		TaskEndDetails:       projCtx.Task.Details,
-		Distro:               projCtx.Task.DistroId,
-		Container:            projCtx.Task.Container,
-		BuildVariant:         projCtx.Task.BuildVariant,
-		BuildId:              projCtx.Task.BuildId,
-		Activated:            projCtx.Task.Activated,
-		Execution:            taskExecution,
-		Requester:            projCtx.Task.Requester,
-		CreateTime:           projCtx.Task.CreateTime,
-		IngestTime:           projCtx.Task.IngestTime,
-		StartTime:            projCtx.Task.StartTime.UnixNano(),
-		DispatchTime:         projCtx.Task.DispatchTime.UnixNano(),
-		FinishTime:           projCtx.Task.FinishTime.UnixNano(),
-		ExpectedDuration:     projCtx.Task.ExpectedDuration,
-		TimeTaken:            projCtx.Task.TimeTaken,
-		Priority:             projCtx.Task.Priority,
-		Aborted:              projCtx.Task.Aborted,
-		AbortInfo:            projCtx.Task.AbortInfo,
-		DisplayOnly:          projCtx.Task.DisplayOnly,
-		OverrideDependencies: projCtx.Task.OverrideDependencies,
-		CurrentTime:          time.Now().UnixNano(),
-		BuildVariantDisplay:  projCtx.Build.DisplayName,
-		Message:              projCtx.Version.Message,
-		Project:              projCtx.Version.Identifier,
-		Author:               projCtx.Version.Author,
-		AuthorEmail:          projCtx.Version.AuthorEmail,
-		VersionId:            projCtx.Version.Id,
-		RepoOwner:            projCtx.ProjectRef.Owner,
-		Repo:                 projCtx.ProjectRef.Repo,
-		Archived:             archived,
-		TotalExecutions:      totalExecutions,
-		PartOfDisplay:        projCtx.Task.IsPartOfDisplay(r.Context()),
-		GeneratedById:        projCtx.Task.GeneratedBy,
-	}
-
-	deps, taskWaiting, err := getTaskDependencies(r.Context(), projCtx.Task)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	uiTask.DependsOn = deps
-	uiTask.TaskWaiting = taskWaiting
-	uiTask.MinQueuePos, err = model.FindMinimumQueuePositionForTask(ctx, uiTask.Id)
-	if err != nil {
-		uis.LoggedError(w, r, http.StatusInternalServerError, err)
-		return
-	}
-	if uiTask.MinQueuePos < 0 {
-		uiTask.MinQueuePos = 0
-	}
-	if uiTask.Status == evergreen.TaskUndispatched {
-		uiTask.EstWaitTime, err = model.GetEstimatedStartTime(ctx, *projCtx.Task)
-		if err != nil {
-			uis.LoggedError(w, r, http.StatusInternalServerError, err)
-			return
-		}
-	}
-	if uiTask.GeneratedById != "" {
-		var generator *task.Task
-		generator, err = task.FindOneIdWithFields(ctx, uiTask.GeneratedById, task.DisplayNameKey)
-		if err != nil {
-			uis.LoggedError(w, r, http.StatusInternalServerError, err)
-			return
-		}
-		uiTask.GeneratedByName = generator.DisplayName
-	}
-
-	var taskHost *host.Host
-	if projCtx.Task.HostId != "" {
-		uiTask.HostDNS = projCtx.Task.HostId
-		uiTask.HostId = projCtx.Task.HostId
-		taskHost, err = host.FindOne(ctx, host.ById(projCtx.Task.HostId))
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-		if taskHost != nil {
-			uiTask.HostDNS = taskHost.Host
-			// ensure that the ability to spawn is updated from the existing distro
-			taskHost.Distro.SpawnAllowed = false
-			var d *distro.Distro
-			d, err = distro.FindOneId(ctx, taskHost.Distro.Id)
-			if err != nil {
-				http.Error(w, err.Error(), http.StatusInternalServerError)
-				return
-			}
-			if d != nil {
-				taskHost.Distro.SpawnAllowed = d.SpawnAllowed
-			}
-		}
-	}
-
-	testResults := uis.getTestResults(r.Context(), projCtx, &uiTask)
-	if projCtx.Patch != nil {
-		var taskOnBaseCommit *task.Task
-		var testResultsOnBaseCommit []testresult.TestResult
-		taskOnBaseCommit, err = projCtx.Task.FindTaskOnBaseCommit(r.Context())
-		if err != nil {
-			uis.LoggedError(w, r, http.StatusInternalServerError, err)
-			return
-		}
-		taskPatch := &uiPatch{Patch: *projCtx.Patch}
-		if taskOnBaseCommit != nil {
-			if err = taskOnBaseCommit.PopulateTestResults(r.Context()); err != nil {
-				uis.LoggedError(w, r, http.StatusInternalServerError, err)
-				return
-			}
-
-			taskPatch.BaseTaskId = taskOnBaseCommit.Id
-			taskPatch.BaseTimeTaken = taskOnBaseCommit.TimeTaken
-			testResultsOnBaseCommit = taskOnBaseCommit.LocalTestResults
-		}
-		taskPatch.StatusDiffs = model.StatusDiffTests(testResultsOnBaseCommit, testResults)
-		uiTask.PatchInfo = taskPatch
-	}
-
-	if projCtx.Task.TriggerID != "" {
-		var projectName string
-		projectName, err = model.GetUpstreamProjectName(r.Context(), projCtx.Task.TriggerID, projCtx.Task.TriggerType)
-		if err != nil {
-			uis.LoggedError(w, r, http.StatusInternalServerError, err)
-			return
-		}
-		uiTask.UpstreamData = &uiUpstreamData{
-			ProjectName: projectName,
-			TriggerID:   projCtx.Task.TriggerID,
-			TriggerType: projCtx.Task.TriggerType,
-		}
-	}
-
-	usr := gimlet.GetUser(ctx)
-	pluginContext := projCtx.ToPluginContext(uis.Settings, usr)
-	pluginContext.Request = r
-	pluginContent := getPluginDataAndHTML(uis, plugin.TaskPage, pluginContext)
-	permissions := gimlet.Permissions{}
-	if usr != nil {
-		opts := gimlet.PermissionOpts{Resource: projCtx.ProjectRef.Id, ResourceType: evergreen.ProjectResourceType}
-		permissions, err = rolemanager.HighestPermissionsForRoles(usr.Roles(), evergreen.GetEnvironment().RoleManager(), opts)
-		if err != nil {
-			uis.LoggedError(w, r, http.StatusInternalServerError, err)
-			return
-		}
-	}
-	newUILink := ""
-	if len(uis.Settings.Ui.UIv2Url) > 0 {
-		newUILink = fmt.Sprintf("%s/task/%s?execution=%d", uis.Settings.Ui.UIv2Url, tId, taskExecution)
-	}
-
-	if uiTask.AbortInfo.TaskID != "" {
-		abortedBy, err := getAbortedBy(r.Context(), projCtx.Task.AbortInfo.TaskID)
-		if err != nil {
-			uis.LoggedError(w, r, http.StatusInternalServerError, err)
-			return
-		}
-		uiTask.AbortedByDisplay = abortedBy
-	}
-
-	uis.render.WriteResponse(w, http.StatusOK, struct {
-		EvgBaseUrl    string
-		Task          uiTaskData
-		Host          *host.Host
-		PluginContent pluginData
-		JiraHost      string
-		Permissions   gimlet.Permissions
-		NewUILink     string
-		ViewData
-	}{uis.Settings.Ui.Url, uiTask, taskHost, pluginContent, uis.Settings.Jira.Host, permissions, newUILink, uis.GetCommonViewData(w, r, false, true)}, "base", "task.html", "base_angular.html", "menu.html")
-}
-
-func getAbortedBy(ctx context.Context, abortedByTaskId string) (*abortedByDisplay, error) {
-	abortedTask, err := task.FindOneId(ctx, abortedByTaskId)
-	if err != nil {
-		return nil, errors.Wrap(err, "problem getting abortedBy task")
-	}
-	buildDisplay, err := build.FindOne(ctx, build.ById(abortedTask.BuildId))
-	if err != nil {
-		return nil, errors.Wrap(err, "problem getting abortedBy build")
-	}
-	if buildDisplay == nil {
-		return nil, errors.New("problem getting abortBy display information")
-	}
-	abortedBy := &abortedByDisplay{
-		TaskDisplayName:     abortedTask.DisplayName,
-		BuildVariantDisplay: buildDisplay.DisplayName,
-	}
-
-	return abortedBy, nil
-}
-
-type taskHistoryPageData struct {
-	TaskName    string
-	Tasks       []bson.M
-	Variants    []string
-	FailedTests map[string][]string
-	Versions    []model.Version
-
-	// Flags that indicate whether the beginning/end of history has been reached
-	ExhaustedBefore bool
-	ExhaustedAfter  bool
-
-	// The revision for which the surrounding history was requested
-	SelectedRevision string
+	http.Redirect(w, r, spruceLink, http.StatusPermanentRedirect)
 }
 
 // the task's most recent log messages
 const DefaultLogMessages = 100 // passed as a limit, so 0 means don't limit
-
-// getTaskDependencies returns the uiDeps for the task and its status (either its original status,
-// "blocked", or "pending")
-func getTaskDependencies(ctx context.Context, t *task.Task) ([]uiDep, string, error) {
-	depIds := []string{}
-	for _, dep := range t.DependsOn {
-		depIds = append(depIds, dep.TaskId)
-	}
-	dependencies, err := task.FindWithFields(ctx, task.ByIds(depIds), task.DisplayNameKey, task.StatusKey,
-		task.ActivatedKey, task.BuildVariantKey, task.DetailsKey, task.DependsOnKey)
-	if err != nil {
-		return nil, "", err
-	}
-	taskMap := map[string]*task.Task{}
-	for i := range dependencies {
-		taskMap[dependencies[i].Id] = &dependencies[i]
-	}
-
-	uiDependencies := []uiDep{}
-	for _, dep := range t.DependsOn {
-		depTask, ok := taskMap[dep.TaskId]
-		if !ok {
-			continue
-		}
-		uiDependencies = append(uiDependencies, uiDep{
-			Id:             dep.TaskId,
-			Name:           depTask.DisplayName,
-			Status:         depTask.Status,
-			RequiredStatus: dep.Status,
-			Activated:      depTask.Activated,
-			BuildVariant:   depTask.BuildVariant,
-			Details:        depTask.Details,
-		})
-	}
-
-	if err = t.CircularDependencies(ctx); err != nil {
-		return nil, "", err
-	}
-	state, err := t.BlockedState(taskMap)
-	if err != nil {
-		return nil, "", errors.Wrap(err, "can't get blocked state")
-	}
-
-	return uiDependencies, state, nil
-}
 
 func (uis *UIServer) taskLog(w http.ResponseWriter, r *http.Request) {
 	projCtx := MustHaveProjectContext(r)
@@ -901,82 +450,4 @@ func (uis *UIServer) testLog(w http.ResponseWriter, r *http.Request) {
 		}
 		uis.render.Stream(w, http.StatusOK, data, "base", "task_log.html")
 	}
-}
-
-func (uis *UIServer) getTestResults(ctx context.Context, projCtx projectContext, uiTask *uiTaskData) []testresult.TestResult {
-	if err := projCtx.Task.PopulateTestResults(ctx); err != nil {
-		grip.Error(message.WrapError(err, message.Fields{
-			"task_id": projCtx.Task.Id,
-			"message": "fetching test results for task",
-		}))
-		return nil
-	}
-
-	uiTask.TestResults = []uiTestResult{}
-	if uiTask.DisplayOnly {
-		execTaskDisplayNameMap := map[string]string{}
-		for _, t := range projCtx.Task.ExecutionTasks {
-			var (
-				et  *task.Task
-				err error
-			)
-			if uiTask.Archived {
-				et, err = task.FindOneOldByIdAndExecution(ctx, t, projCtx.Task.Execution)
-			} else {
-				et, err = task.FindOneId(ctx, t)
-			}
-			if err != nil {
-				grip.Error(message.Fields{
-					"message":        "fetching test results for execution task",
-					"task_id":        t,
-					"parent_task_id": projCtx.Task.Id,
-				})
-				return nil
-			}
-			if et == nil {
-				grip.Error(message.Fields{
-					"message":        "execution task not found",
-					"task_id":        t,
-					"parent_task_id": projCtx.Task.Id,
-				})
-				continue
-			}
-
-			execTaskDisplayNameMap[t] = et.DisplayName
-			uiTask.ExecutionTasks = append(uiTask.ExecutionTasks, uiExecTask{
-				Id:        t,
-				Name:      et.DisplayName,
-				TimeTaken: et.TimeTaken,
-				Status:    et.ResultStatus(),
-			})
-		}
-
-		for _, tr := range projCtx.Task.LocalTestResults {
-			uiTask.TestResults = append(uiTask.TestResults, uiTestResult{
-				TestResult: tr,
-				TaskId:     tr.TaskID,
-				TaskName:   execTaskDisplayNameMap[tr.TaskID],
-				URL:        tr.GetLogURL(uis.env, evergreen.LogViewerHTML),
-				URLRaw:     tr.GetLogURL(uis.env, evergreen.LogViewerRaw),
-				URLParsley: tr.GetLogURL(uis.env, evergreen.LogViewerParsley),
-			})
-		}
-	} else {
-		for _, tr := range projCtx.Context.Task.LocalTestResults {
-			uiTask.TestResults = append(uiTask.TestResults, uiTestResult{
-				TestResult: tr,
-				TaskId:     tr.TaskID,
-				URL:        tr.GetLogURL(uis.env, evergreen.LogViewerHTML),
-				URLRaw:     tr.GetLogURL(uis.env, evergreen.LogViewerRaw),
-				URLParsley: tr.GetLogURL(uis.env, evergreen.LogViewerParsley),
-			})
-		}
-
-		if uiTask.PartOfDisplay {
-			// Display task ID would've been populated when setting PartOfDisplay.
-			uiTask.DisplayTaskID = utility.FromStringPtr(projCtx.Task.DisplayTaskId)
-		}
-	}
-
-	return projCtx.Task.LocalTestResults
 }
