+++ service/templates/admin.html
@@ -466,15 +466,6 @@ Admin Settings
 											</tr>
 											<tr>
 											<tr>
-												<td>Legacy UI Task Page</td>
-												<td colspan="2">
-													<md-radio-group data-ng-model="Settings.service_flags.legacy_ui_task_page_disabled" layout="row">
-														<md-radio-button data-ng-value="false"></md-radio-button>
-														<md-radio-button data-ng-value="true"></md-radio-button>
-													</md-radio-group>
-												</td>
-											</tr>
-											<tr>
 												<td>Legacy UI Task History Page</td>
 												<td colspan="2">
 													<md-radio-group data-ng-model="Settings.service_flags.legacy_ui_task_history_page_disabled" layout="row">
