+++ testutil/config.go
@@ -398,7 +398,6 @@ func MockConfig() *evergreen.Settings {
 			ElasticIPsDisabled:              true,
 			ReleaseModeDisabled:             true,
 			AdminParameterStoreDisabled:     false,
-			LegacyUITaskPageDisabled:        true,
 			LegacyUITaskHistoryPageDisabled: true,
 		},
 		SingleTaskDistro: evergreen.SingleTaskDistroConfig{
